requirements.txt
openai
langchain
langchain-community
langchain-openai
serpapi
google-search-results
gspread
python-dotenv
